refactors to create deviceshadowhandler.py
